editoast: move middleware definitions in `server::middlewares`
De-clutters `views/mod.rs`.
